requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
baidu-aip==2.2.0.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.1
cssselect==1.0.3
Django==2.0.2
django-celery-results==1.0.1
django-redis==4.9.0
djangorestframework==3.7.7
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
kombu==4.1.0
lxml==4.2.1
mysql-connector-python==8.0.6
optionaldict==0.1.1
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2018.3
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
vine==1.1.4
w3lib==1.19.0
xmltodict==0.11.0
zope.interface==4.4.3
